[dev]
Created a shell script which will be able to run the simulator and scaling manager in a accelerated way.

As of now the acceleration is possible for one day. The acceleration factor is following:
1 seconds = 5 minute
To Do:
In future we should be able to simulate it for whole week or month based on criteria.