Keep Alive on HTTP connection is disabled.
The server's default keep alive time is 2 hours. This keeps your http socket connection active for 2 hours even if you don't utilise it.
This opens a socket for each request made via the opensearch client.
As a result, there are too many open connections at any given time.

This client setting ends the http connection after the response to that request is served.